app/requirement.txt
click==7.1.2
DAWG-Python==0.7.2
docopt==0.6.2
fastapi==0.63.0
h11==0.12.0
inexactsearch==1.0.2
intervaltree==3.1.0
ipymarkup==0.9.0
natasha==1.4.0
navec==0.10.0
numpy==1.20.0
pydantic==1.7.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyspellchecker==0.5.6
razdel==0.5.0
silpa-common==0.3
six==1.15.0
slovnet==0.5.0
sortedcontainers==2.3.0
soundex==1.1.3
starlette==0.13.6
transliterate==1.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
yargy==0.15.0
